requirements.txt
psychopy>3.0.4
pyyaml
pandas>0.23.0
numpy>1.14.3
msgpack_numpy
